Introduce qwen3_tts/prompt_embeds_builder
Introduce a reusable prompt embeds builder for qwen3tts,
so it can be reused by other alternative implementations
