+++ vllm_omni/model_executor/models/qwen3_tts/qwen3_tts_talker.py
@@ -1,21 +1,14 @@
 from __future__ import annotations
 
-import base64
 import copy
-import hashlib
-import io
-from collections import OrderedDict
-from collections.abc import Callable, Iterable, Mapping
-from functools import lru_cache
+from collections.abc import Iterable, Mapping
 from typing import Any
-from urllib.parse import urlparse
 
 import numpy as np
-import soundfile as sf
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-from transformers import AutoFeatureExtractor, AutoTokenizer
+from transformers import AutoFeatureExtractor
 from transformers.activations import ACT2FN
 from vllm.config import VllmConfig
 from vllm.distributed import get_pp_group
@@ -30,10 +23,10 @@ from vllm.sequence import IntermediateTensors
 
 from vllm_omni.data_entry_keys import OmniPayload
 from vllm_omni.model_executor.models.output_templates import OmniOutput
-from vllm_omni.utils.audio import mel_filter_bank
 from vllm_omni.utils.speaker_cache import get_speaker_cache
 
 from .configuration_qwen3_tts import Qwen3TTSConfig, Qwen3TTSSpeakerEncoderConfig, Qwen3TTSTalkerConfig
+from .prompt_embeds_builder import Qwen3TTSPromptEmbedsBuilder
 from .qwen3_tts_code_predictor_vllm import Qwen3TTSTalkerCodePredictorForConditionalGenerationVLLM
 from .tokenizer_12hz.configuration_qwen3_tts_tokenizer_v2 import Qwen3TTSTokenizerV2Config
 from .tokenizer_12hz.modeling_qwen3_tts_tokenizer_v2 import Qwen3TTSTokenizerV2Encoder
@@ -41,11 +34,6 @@ from .tokenizer_12hz.modeling_qwen3_tts_tokenizer_v2 import Qwen3TTSTokenizerV2E
 logger = init_logger(__name__)
 
 _TRAILING_TEXT_COMPACT_MIN_FRAMES = 64
-_PRECOMPUTED_REF_CODE_KEY = "precomputed_ref"
-_NORMALIZED_REF_AUDIO_KEY = "_qwen3_tts_normalized_ref_audio"
-_REF_AUDIO_CACHE_KEY = "_qwen3_tts_ref_audio_cache_key"
-_PRECOMPUTED_TEXT_IDS_KEY = "_qwen3_tts_text_ids"
-_PRECOMPUTED_REF_IDS_KEY = "_qwen3_tts_ref_ids"
 
 
 # ---------------------------------------------------------------------------
@@ -252,69 +240,6 @@ class Qwen3TTSSpeakerEncoder(torch.nn.Module):
         return hidden_states.squeeze(-1)
 
 
-# ---- Audio utilities ----
-
-
-def _dynamic_range_compression(x, c=1, clip_val=1e-5):
-    return torch.log(torch.clamp(x, min=clip_val) * c)
-
-
-@lru_cache(maxsize=8)
-def _cached_mel_filter_bank(sampling_rate: int, n_fft: int, n_mels: int, fmin: int, fmax: int | None) -> torch.Tensor:
-    return mel_filter_bank(sr=sampling_rate, n_fft=n_fft, n_mels=n_mels, fmin=fmin, fmax=fmax)
-
-
-@lru_cache(maxsize=8)
-def _cached_hann_window(win_size: int) -> torch.Tensor:
-    return torch.hann_window(win_size)
-
-
-def mel_spectrogram(
-    y: torch.Tensor,
-    n_fft: int,
-    num_mels: int,
-    sampling_rate: int,
-    hop_size: int,
-    win_size: int,
-    fmin: int,
-    fmax: int | None = None,
-    center: bool = False,
-    mel_basis: torch.Tensor | None = None,
-    hann_window: torch.Tensor | None = None,
-) -> torch.Tensor:
-    """Calculate mel spectrogram of an input signal using torchaudio mel filterbank and torch STFT."""
-    if torch.min(y) < -1.0:
-        logger.warning("Min value of input waveform signal is %s", torch.min(y))
-    if torch.max(y) > 1.0:
-        logger.warning("Max value of input waveform signal is %s", torch.max(y))
-    device = y.device
-    if mel_basis is None:
-        mel_basis = _cached_mel_filter_bank(sampling_rate, n_fft, num_mels, fmin, fmax).to(device)
-    elif mel_basis.device != device:
-        mel_basis = mel_basis.to(device)
-    if hann_window is None:
-        hann_window = _cached_hann_window(win_size).to(device)
-    elif hann_window.device != device:
-        hann_window = hann_window.to(device)
-    padding = (n_fft - hop_size) // 2
-    y = torch.nn.functional.pad(y.unsqueeze(1), (padding, padding), mode="reflect").squeeze(1)
-    spec = torch.stft(
-        y,
-        n_fft,
-        hop_length=hop_size,
-        win_length=win_size,
-        window=hann_window,
-        center=center,
-        pad_mode="reflect",
-        normalized=False,
-        onesided=True,
-        return_complex=True,
-    )
-    spec = torch.sqrt(torch.view_as_real(spec).pow(2).sum(-1) + 1e-9)
-    mel_spec = torch.matmul(mel_basis, spec)
-    return _dynamic_range_compression(mel_spec)
-
-
 # ---------------------------------------------------------------------------
 # Main AR Talker model
 # ---------------------------------------------------------------------------
@@ -471,9 +396,6 @@ class Qwen3TTSTalkerForConditionalGeneration(nn.Module):
             persistent=False,
         )
 
-        # Tokenizer for prompt building.
-        self._tokenizer = None
-
         tokenizer_config = Qwen3TTSTokenizerV2Config.from_pretrained(
             self.model_path,
             subfolder="speech_tokenizer",
@@ -492,17 +414,33 @@ class Qwen3TTSTalkerForConditionalGeneration(nn.Module):
         )
 
         self._speaker_cache = get_speaker_cache()
-        self._ref_audio_artifact_cache_max_entries = 256
-        self._ref_audio_artifact_cache: OrderedDict[str, dict[str, torch.Tensor | bool]] = OrderedDict()
         raw_subtalker_sampling = getattr(vllm_config.model_config, "subtalker_sampling_params", None)
         self._subtalker_sampling_params: dict[str, Any] = (
             dict(raw_subtalker_sampling) if isinstance(raw_subtalker_sampling, Mapping) else {}
         )
 
         self._stacked_codec_embed: torch.Tensor | None = None
-        # Bounded LRU: caller-supplied orig_sr can otherwise grow this without limit.
-        self._resampler_cache: OrderedDict[tuple[int, int], AudioResampler] = OrderedDict()
-        self._resampler_cache_max = 16
+
+        # Stand-alone wrapper around the embedding layers, encoders and
+        # tokenizers required to assemble a talker prefill prompt. Owns the
+        # text + speech tokenizers (loaded lazily on first use) and the
+        # per-ref-audio / resampler caches, so they don't leak onto the
+        # talker class. Other talker variants can construct their own builder
+        # with the same set of dependencies and reuse ``build_prompt_embeds``
+        # verbatim.
+        self._prompt_builder = Qwen3TTSPromptEmbedsBuilder(
+            config=self.config,
+            talker_config=self.talker_config,
+            model_path=self.model_path,
+            text_embedding=self.text_embedding,
+            text_projection=self.text_projection,
+            codec_embed=self.embed_input_ids,
+            residual_code_embeddings=lambda: self.code_predictor.get_input_embeddings(),
+            speaker_encoder=self.speaker_encoder,
+            tts_pad_embed=self._tts_pad_embed,
+            encode_ref_audio_batch=self._encode_ref_audio_batch,
+            speaker_cache=self._speaker_cache,
+        )
 
     # -------------------- vLLM required hooks --------------------
 
@@ -669,8 +607,8 @@ class Qwen3TTSTalkerForConditionalGeneration(nn.Module):
             # Subsequent prefill rounds (multi-chunk): prompt_embeds_cpu is a Tensor stored by the first round.
             is_first_prefill = not isinstance(prompt_embeds_cpu, torch.Tensor) or prompt_embeds_cpu.ndim != 2
             if is_first_prefill:
-                full_prompt_embeds, tailing_text_hidden, ref_code_len, ref_code = self._build_prompt_embeds(
-                    task_type=task_type, info_dict=info_dict
+                full_prompt_embeds, tailing_text_hidden, ref_code_len, ref_code = (
+                    self._prompt_builder.build_prompt_embeds(task_type=task_type, info_dict=info_dict)
                 )
                 # Store full prompt embeddings on CPU (large, prefill-only).
                 # tailing_text_hidden stays on GPU (gpu_resident_buffer_keys).
@@ -900,494 +838,20 @@ class Qwen3TTSTalkerForConditionalGeneration(nn.Module):
         last = hidden_states[-1, :].detach()
         return {"hidden_states": {"last": last}}
 
-    # -------------------- prompt construction helpers --------------------
-
-    def _get_tokenizer(self):
-        if self._tokenizer is None:
-            import transformers
-
-            kwargs = dict(trust_remote_code=True, use_fast=True)
-            if transformers.__version__ < "5":
-                kwargs["fix_mistral_regex"] = True
-            self._tokenizer = AutoTokenizer.from_pretrained(self.model_path, **kwargs)
-            self._tokenizer.padding_side = "left"
-        return self._tokenizer
-
-    @staticmethod
-    def _build_assistant_text(text: str) -> str:
-        return f"<|im_start|>assistant\n{text}<|im_end|>\n<|im_start|>assistant\n"
-
-    @staticmethod
-    def _build_ref_text(text: str) -> str:
-        return f"<|im_start|>assistant\n{text}<|im_end|>\n"
-
-    @staticmethod
-    def _build_instruct_text(instruct: str) -> str:
-        return f"<|im_start|>user\n{instruct}<|im_end|>\n"
-
-    @staticmethod
-    def estimate_prompt_len_from_additional_information(
-        additional_information: dict[str, Any] | None,
-        *,
-        task_type: str,
-        tokenize_prompt: Callable[[str], list[int]],
-        codec_language_id: Mapping[str, int] | None,
-        spk_is_dialect: Mapping[str, object] | None,
-        estimate_ref_code_len: Callable[[object], int | None] | None = None,
-    ) -> int:
-        """Compute Stage-0 placeholder prompt length (length-only mirror of `_build_prompt_embeds()`).
-        It must match the model-side `inputs_embeds` length to avoid extra padding and quality drop."""
-
-        def _first(x: object, default: object) -> object:
-            if isinstance(x, list):
-                return x[0] if x else default
-            return x if x is not None else default
-
-        info: dict[str, Any] = additional_information or {}
-        text = _first(info.get("text"), "")
-        language = _first(info.get("language"), "Auto")
-        speaker = _first(info.get("speaker"), "").lower().strip()
-        instruct = _first(info.get("instruct"), "")
-        non_streaming_mode_raw = _first(info.get("non_streaming_mode"), None)
-
-        if isinstance(non_streaming_mode_raw, bool):
-            non_streaming_mode = non_streaming_mode_raw
-        else:
-            # Official defaults: CustomVoice/VoiceDesign -> non_streaming_mode=True; Base -> False.
-            non_streaming_mode = task_type in ("CustomVoice", "VoiceDesign")
-
-        if not isinstance(text, str):
-            text = ""
-        if not isinstance(instruct, str):
-            instruct = ""
-        if not isinstance(language, str):
-            language = "Auto"
-
-        instruct_len = 0
-        if instruct.strip():
-            instruct_text = Qwen3TTSTalkerForConditionalGeneration._build_instruct_text(instruct)
-            instruct_len = len(tokenize_prompt(instruct_text))
-
-        # ---- codec prefix portion (matches _build_prompt_embeds) ----
-        language_id = None
-        if language.lower() != "auto" and codec_language_id:
-            language_id = codec_language_id.get(language.lower())
-        if (
-            language_id is None
-            and codec_language_id
-            and spk_is_dialect
-            and isinstance(language, str)
-            and language.lower() in ("chinese", "auto")
-            and isinstance(speaker, str)
-            and speaker.strip()
-        ):
-            dialect = spk_is_dialect.get(speaker.lower())
-            if isinstance(dialect, str) and dialect:
-                language_id = codec_language_id.get(dialect)
-        prefill_len = 3 if language_id is None else 4
-
-        speaker_len = 1 if task_type in ("CustomVoice", "Base") else 0
-        codec_input_len = prefill_len + speaker_len + 2  # + [codec_pad, codec_bos]
-        codec_prefix_len = codec_input_len - 1  # codec_input[:-1] + tts_bos
-
-        # Role header: input_ids[:, :3] in model.
-        role_len = 3
-        prompt_len = instruct_len + role_len + codec_prefix_len
-
-        # ---- text conditioning portion (matches _build_prompt_embeds) ----
-        assistant_text = Qwen3TTSTalkerForConditionalGeneration._build_assistant_text(text)
-        assistant_len = len(tokenize_prompt(assistant_text))
-        if assistant_len < 8:
-            raise ValueError(f"Unexpected assistant prompt length: {assistant_len}")
-
-        if task_type in ("CustomVoice", "VoiceDesign"):
-            if non_streaming_mode:
-                # model: full text ids (input_ids[:, 3:-5]) + eos + codec_bos step
-                prompt_len += assistant_len - 6
-            else:
-                # model: only first text token in prefill
-                prompt_len += 1
-
-        if task_type == "Base":
-            xvec_only = bool(_first(info.get("x_vector_only_mode"), False))
-            in_context_mode = not xvec_only
-
-            voice_clone_prompt = _first(info.get("voice_clone_prompt"), None)
-            if isinstance(voice_clone_prompt, dict):
-                icl_flag = _first(voice_clone_prompt.get("icl_mode"), None)
-                if isinstance(icl_flag, bool):
-                    in_context_mode = icl_flag
-
-            if in_context_mode:
-                ref_code = None
-                if isinstance(voice_clone_prompt, dict):
-                    ref_code = _first(voice_clone_prompt.get("ref_code"), None)
-
-                ref_code_len: int | None = None
-                if isinstance(ref_code, list):
-                    if ref_code and isinstance(ref_code[0], list):
-                        ref_code_len = len(ref_code)
-                    elif ref_code:
-                        ref_code_len = len(ref_code)
-                elif hasattr(ref_code, "shape"):
-                    try:
-                        shape = getattr(ref_code, "shape")
-                        if shape and len(shape) >= 1:
-                            ref_code_len = int(shape[0])
-                    except Exception:
-                        ref_code_len = None
-
-                if ref_code_len is None and estimate_ref_code_len is not None:
-                    ref_code_len = estimate_ref_code_len(info.get("ref_audio"))
-                if ref_code_len is None:
-                    raise ValueError(
-                        "Base in-context voice cloning requires either `voice_clone_prompt.ref_code` "
-                        "or a readable `ref_audio` that can be mapped to a codec frame length."
-                    )
-
-                codec_lens = 1 + int(ref_code_len)  # codec_bos + ref_code
-                if non_streaming_mode:
-                    # _generate_icl_prompt(non_streaming_mode=True):
-                    # text_embed = ref_ids + text_ids + eos.
-                    ref_ids = _first(info.get("ref_ids"), None)
-                    if isinstance(voice_clone_prompt, dict) and ref_ids is None:
-                        ref_ids = _first(voice_clone_prompt.get("ref_ids") or voice_clone_prompt.get("ref_id"), None)
-
-                    if ref_ids is None:
-                        ref_text = _first(info.get("ref_text"), "")
-                        if not isinstance(ref_text, str) or not ref_text.strip():
-                            raise ValueError(
-                                "Base in-context non-streaming requires `ref_text` or tokenized `ref_ids`."
-                            )
-                        ref_text_ids = tokenize_prompt(Qwen3TTSTalkerForConditionalGeneration._build_ref_text(ref_text))
-                        ref_ids_len = len(ref_text_ids)
-                    elif hasattr(ref_ids, "shape"):
-                        shape = getattr(ref_ids, "shape", None)
-                        ref_ids_len = int(shape[-1]) if shape else 0
-                    elif isinstance(ref_ids, list):
-                        ref_ids_len = len(ref_ids)
-                    else:
-                        ref_ids_len = 0
-
-                    # model uses ref_ids[:, 3:-2] (strip 5 tokens) and text_id=input_ids[:, 3:-5] (strip 8).
-                    ref_id_len = max(0, int(ref_ids_len) - 5)
-                    text_id_len = max(0, int(assistant_len) - 8)
-                    text_embed_len = ref_id_len + text_id_len + 1  # + eos
-                    prompt_len += text_embed_len + codec_lens
-                else:
-                    # _generate_icl_prompt(non_streaming_mode=False): aligned to codec_lens.
-                    prompt_len += codec_lens
-            else:
-                # Base without ICL behaves like CustomVoice.
-                if non_streaming_mode:
-                    prompt_len += assistant_len - 6
-                else:
-                    prompt_len += 1
-
-        return max(2, int(prompt_len))
-
-    def _is_probably_base64(self, s: str) -> bool:
-        if s.startswith("data:audio"):
-            return True
-        if ("/" not in s and "\\" not in s) and len(s) > 256:
-            return True
-        return False
-
-    def _is_url(self, s: str) -> bool:
-        try:
-            u = urlparse(s)
-            if u.scheme in ("http", "https"):
-                return bool(u.netloc)
-            return u.scheme == "file"
-        except Exception:
-            return False
-
-    def _decode_base64_to_wav_bytes(self, b64: str) -> bytes:
-        if "," in b64 and b64.strip().startswith("data:"):
-            b64 = b64.split(",", 1)[1]
-        return base64.b64decode(b64)
-
-    def _load_audio_to_np(self, x: str) -> tuple[np.ndarray, int]:
-        """Load audio from local path, URL, or base64 data URI.
-
-        Uses upstream vLLM's MediaConnector for http(s) URLs and ``file:``
-        URIs, with unrestricted local access (offline inference is trusted).
-        """
-        from vllm.multimodal.media.audio import load_audio
-
-        if self._is_url(x):
-            from vllm.multimodal.media import MediaConnector
-
-            connector = MediaConnector(allowed_local_media_path="/")
-            audio, sr = connector.fetch_audio(x)
-        elif self._is_probably_base64(x):
-            wav_bytes = self._decode_base64_to_wav_bytes(x)
-            with io.BytesIO(wav_bytes) as f:
-                audio, sr = sf.read(f, dtype="float32", always_2d=False)
-        else:
-            audio, sr = load_audio(x, sr=None, mono=True)
-
-        if isinstance(audio, np.ndarray) and audio.ndim > 1:
-            audio = np.mean(audio, axis=-1)
-
-        return np.asarray(audio, dtype=np.float32), int(sr)
-
-    def _normalize_ref_audio(self, ref_audio: object) -> tuple[np.ndarray, int]:
-        # NOTE: additional_information may serialize (wav, sr) into (nested) lists across processes; be tolerant.
-        if isinstance(ref_audio, str):
-            return self._load_audio_to_np(ref_audio)
-
-        def _is_sr(x: object) -> bool:
-            try:
-                v = int(x)  # type: ignore[arg-type]
-            except Exception:
-                return False
-            return 1_000 <= v <= 200_000
-
-        def _is_number_sequence(xs: list[object]) -> bool:
-            if not xs:
-                return False
-            for v in xs[:8]:
-                if not isinstance(v, (int, float, np.number)):
-                    return False
-            return True
-
-        wav_candidates: list[object] = []
-        sr_candidates: list[int] = []
-
-        def _summarize(obj: object, depth: int = 0) -> str:
-            if depth > 2:
-                if isinstance(obj, (int, np.integer)):
-                    return f"int({int(obj)})"
-                return type(obj).__name__
-            if obj is None:
-                return "None"
-            if isinstance(obj, str):
-                if len(obj) <= 16:
-                    return f"str({obj!r})"
-                return f"str(len={len(obj)})"
-            if isinstance(obj, (int, float, np.number)):
-                return f"{type(obj).__name__}({obj})"
-            if isinstance(obj, np.ndarray):
-                return f"ndarray(shape={obj.shape}, dtype={obj.dtype})"
-            if isinstance(obj, torch.Tensor):
-                return f"Tensor(shape={tuple(obj.shape)}, dtype={obj.dtype}, device={obj.device})"
-            if isinstance(obj, dict):
-                keys = list(obj.keys())
-                return f"dict(keys={keys[:8]})"
-            if isinstance(obj, (tuple, list)):
-                items = list(obj)
-                head = ", ".join(_summarize(x, depth + 1) for x in items[:3])
-                return f"{type(obj).__name__}(len={len(items)}; head=[{head}])"
-            return f"{type(obj).__name__}"
-
-        def _scan(obj: object, depth: int = 0) -> None:
-            if depth > 4:
-                return
-            if obj is None:
-                return
-            if _is_sr(obj):
-                sr_candidates.append(int(obj))  # type: ignore[arg-type]
-                return
-            if isinstance(obj, np.ndarray) and obj.size > 0:
-                wav_candidates.append(obj)
-                return
-            if isinstance(obj, torch.Tensor) and obj.numel() > 0:
-                wav_candidates.append(obj)
-                return
-            if isinstance(obj, dict):
-                # Inlined ndarray/tensor payloads from the input processor.
-                if obj.get("__ndarray__") and "data" in obj and "dtype" in obj and "shape" in obj:
-                    try:
-                        data = obj["data"]
-                        dtype = obj["dtype"]
-                        shape = obj["shape"]
-                        if isinstance(data, (bytes, bytearray, memoryview)):
-                            arr = np.frombuffer(data, dtype=dtype).reshape(shape)
-                            if arr.size > 0:
-                                wav_candidates.append(arr)
-                                return
-                    except Exception:
-                        pass
-                if obj.get("__tensor__") and "data" in obj and "dtype" in obj and "shape" in obj:
-                    try:
-                        data = obj["data"]
-                        dtype = obj["dtype"]
-                        shape = obj["shape"]
-                        if isinstance(data, (bytes, bytearray, memoryview)):
-                            # Stored as raw CPU bytes; interpret as numpy for audio.
-                            np_dtype = np.dtype(dtype)
-                            arr = np.frombuffer(data, dtype=np_dtype).reshape(shape)
-                            if arr.size > 0:
-                                wav_candidates.append(arr)
-                                return
-                    except Exception:
-                        pass
-                wav_obj = obj.get("array") or obj.get("wav") or obj.get("audio")
-                sr_obj = obj.get("sampling_rate") or obj.get("sr") or obj.get("sample_rate")
-                if wav_obj is not None:
-                    _scan(wav_obj, depth + 1)
-                if sr_obj is not None:
-                    _scan(sr_obj, depth + 1)
-                return
-            if isinstance(obj, (tuple, list)):
-                obj_list = list(obj)
-                # Unwrap singleton nesting ([[wav, sr]]).
-                while isinstance(obj_list, list) and len(obj_list) == 1:
-                    inner = obj_list[0]
-                    if isinstance(inner, np.ndarray) and inner.size > 0:
-                        wav_candidates.append(inner)
-                        return
-                    if isinstance(inner, torch.Tensor) and inner.numel() > 0:
-                        wav_candidates.append(inner)
-                        return
-                    if isinstance(inner, dict):
-                        _scan(inner, depth + 1)
-                        return
-                    if isinstance(inner, (tuple, list)):
-                        obj_list = list(inner)  # type: ignore[list-item]
-                        continue
-                    break
-
-                # If the *unwrapped* list is a long list of numbers, treat it as waveform.
-                if len(obj_list) >= 512 and _is_number_sequence(obj_list):
-                    wav_candidates.append(obj_list)
-                    return
-
-                # Otherwise, recurse into elements (but avoid descending into huge numeric lists).
-                for item in obj_list:
-                    if isinstance(item, list) and len(item) >= 512 and _is_number_sequence(item):  # type: ignore[arg-type]
-                        wav_candidates.append(item)
-                        continue
-                    _scan(item, depth + 1)
-                return
-
-        _scan(ref_audio)
-        if not sr_candidates:
-            raise TypeError(f"ref_audio missing sample_rate: {_summarize(ref_audio)}")
-        sr = int(sr_candidates[0])
-
-        def _wav_len(x: object) -> int:
-            try:
-                if isinstance(x, np.ndarray):
-                    return int(x.size)
-                if isinstance(x, torch.Tensor):
-                    return int(x.numel())
-                if isinstance(x, list):
-                    return int(len(x))
-            except Exception:
-                pass
-            return 0
-
-        if not wav_candidates:
-            raise TypeError(f"ref_audio missing waveform: {_summarize(ref_audio)}")
-        wav_obj = max(wav_candidates, key=_wav_len)
-
-        def _to_np(x: object) -> np.ndarray:
-            if isinstance(x, np.ndarray):
-                return x.astype(np.float32).reshape(-1)
-            if isinstance(x, torch.Tensor):
-                return x.detach().to("cpu").float().contiguous().numpy().reshape(-1)
-            if isinstance(x, dict) and x.get("__ndarray__") and "data" in x and "dtype" in x and "shape" in x:
-                data = x["data"]
-                dtype = x["dtype"]
-                shape = x["shape"]
-                if isinstance(data, (bytes, bytearray, memoryview)):
-                    return np.frombuffer(data, dtype=dtype).reshape(shape).astype(np.float32).reshape(-1)
-            if isinstance(x, list):
-                # list of numbers
-                if len(x) >= 2 and _is_number_sequence(x):  # type: ignore[arg-type]
-                    return np.asarray(x, dtype=np.float32).reshape(-1)
-                # list of chunks
-                parts: list[np.ndarray] = []
-                for part in x:
-                    if isinstance(part, (np.ndarray, torch.Tensor, list)):
-                        parts.append(_to_np(part))
-                if parts:
-                    return np.concatenate(parts, axis=0)
-            raise TypeError(f"Unsupported waveform type: {type(x)}")
-
-        wav_np = _to_np(wav_obj)
-        if wav_np.size < 1024:
-            raise ValueError(f"ref_audio waveform too short: {wav_np.size} samples")
-        return wav_np, sr
-
-    def _extract_speaker_embedding(self, wav: np.ndarray, sr: int) -> torch.Tensor:
-        # vLLM workers do not automatically move arbitrary torch.nn.Modules to
-        # CUDA. Ensure the speaker encoder is on the same device/dtype as the
-        # main model before running it.
-        dev = next(self.parameters()).device
-        try:
-            spk_param = next(self.speaker_encoder.parameters())
-            if spk_param.device != dev or spk_param.dtype != torch.bfloat16:
-                self.speaker_encoder.to(device=dev, dtype=torch.bfloat16)
-        except StopIteration:
-            pass
-        # Resample to 24kHz for speaker encoder.
-        target_sr = int(getattr(self.config.speaker_encoder_config, "sample_rate", 24000))
-        if sr != target_sr:
-            resampler = self._get_resampler(int(sr), target_sr)
-            wav = resampler.resample(wav.astype(np.float32), orig_sr=int(sr))
-            sr = target_sr
-
-        # Follow official implementation: mel_spectrogram expects 24kHz.  Move
-        # the waveform first so STFT/mel computation stays on the model device
-        # instead of materializing a CPU mel tensor and copying it per request.
-        wav_tensor = torch.from_numpy(wav).to(device=dev, dtype=torch.float32).unsqueeze(0)
-        mel_basis, hann_window = self._get_speaker_mel_buffers(
-            device=dev,
-            sampling_rate=24000,
-            n_fft=1024,
-            num_mels=128,
-            win_size=1024,
-            fmin=0,
-            fmax=12000,
-        )
-        mels = mel_spectrogram(
-            wav_tensor,
-            n_fft=1024,
-            num_mels=128,
-            sampling_rate=24000,
-            hop_size=256,
-            win_size=1024,
-            fmin=0,
-            fmax=12000,
-            mel_basis=mel_basis,
-            hann_window=hann_window,
-        ).transpose(1, 2)
-        spk = self.speaker_encoder(mels.to(dtype=torch.bfloat16))[0]
-        return spk.to(dtype=torch.bfloat16)
-
-    def _get_speaker_mel_buffers(
+    @torch.inference_mode()
+    def preprocess_batch(
         self,
         *,
+        req_ids: list[str],
+        model_intermediate_buffer: dict[str, dict[str, Any]],
         device: torch.device,
-        sampling_rate: int,
-        n_fft: int,
-        num_mels: int,
-        win_size: int,
-        fmin: int,
-        fmax: int | None,
-    ) -> tuple[torch.Tensor, torch.Tensor]:
-        cache = getattr(self, "_speaker_mel_buffer_cache", None)
-        if cache is None:
-            cache = {}
-            self._speaker_mel_buffer_cache = cache
-        key = (str(device), int(sampling_rate), int(n_fft), int(num_mels), int(win_size), int(fmin), fmax)
-        cached = cache.get(key)
-        if cached is not None:
-            return cached
-        mel_basis = mel_filter_bank(
-            sr=int(sampling_rate),
-            n_fft=int(n_fft),
-            n_mels=int(num_mels),
-            fmin=int(fmin),
-            fmax=fmax,
-        ).to(device)
-        hann_window = torch.hann_window(int(win_size), device=device)
-        cached = (mel_basis, hann_window)
-        cache[key] = cached
-        return cached
+    ) -> None:
+        """Delegate batched preprocess to :class:`Qwen3TTSPromptEmbedsBuilder`."""
+        self._prompt_builder.preprocess_batch(
+            req_ids=req_ids,
+            model_intermediate_buffer=model_intermediate_buffer,
+            device=device,
+        )
 
     def _encode_ref_audio_batch(self, wavs: list[np.ndarray], sr: int, *, device: torch.device) -> list[torch.Tensor]:
         fe = self._encoder_feature_extractor
@@ -1418,765 +882,6 @@ class Qwen3TTSTalkerForConditionalGeneration(nn.Module):
             code[..., : -(-mask.sum() // downsample)].transpose(0, 1).to(device=device, dtype=torch.long)
             for code, mask in zip(audio_codes, padding_mask)
         ]
-
-    @staticmethod
-    def _make_ref_audio_cache_key(wav: np.ndarray, sr: int) -> str:
-        wav_f32 = wav.astype(np.float32, copy=False).reshape(-1)
-        h = hashlib.sha1()
-        h.update(int(sr).to_bytes(4, byteorder="little", signed=False))
-        h.update(int(wav_f32.size).to_bytes(8, byteorder="little", signed=False))
-        h.update(wav_f32.tobytes(order="C"))
-        return h.hexdigest()
-
-    def _get_ref_audio_artifacts(self, cache_key: str) -> dict[str, torch.Tensor | bool] | None:
-        entry = self._ref_audio_artifact_cache.get(cache_key)
-        if entry is None:
-            return None
-        # Touch for LRU.
-        self._ref_audio_artifact_cache.move_to_end(cache_key, last=True)
-        return entry
-
-    def _put_ref_audio_artifacts(
-        self,
-        cache_key: str,
-        *,
-        ref_code: torch.Tensor | None = None,
-        ref_spk_embedding: torch.Tensor | None = None,
-    ) -> None:
-        if not cache_key:
-            return
-        entry = self._ref_audio_artifact_cache.get(cache_key)
-        if entry is None:
-            entry = {}
-        if isinstance(ref_code, torch.Tensor):
-            entry["ref_code"] = ref_code.detach().to("cpu", dtype=torch.long).contiguous()
-        if isinstance(ref_spk_embedding, torch.Tensor):
-            entry["ref_spk_embedding"] = ref_spk_embedding.detach().to("cpu", dtype=torch.bfloat16).reshape(-1)
-        if not entry:
-            return
-        self._ref_audio_artifact_cache[cache_key] = entry
-        self._ref_audio_artifact_cache.move_to_end(cache_key, last=True)
-        while len(self._ref_audio_artifact_cache) > self._ref_audio_artifact_cache_max_entries:
-            self._ref_audio_artifact_cache.popitem(last=False)
-
-    @staticmethod
-    def _first_value(value: object, default: object = None) -> object:
-        if isinstance(value, list):
-            return value[0] if value else default
-        return value if value is not None else default
-
-    @staticmethod
-    def _coerce_ref_code_tensor(value: object, *, device: torch.device) -> torch.Tensor | None:
-        value = Qwen3TTSTalkerForConditionalGeneration._first_value(value)
-        if value is None:
-            return None
-        if isinstance(value, torch.Tensor):
-            ref_code = value
-        elif isinstance(value, np.ndarray):
-            ref_code = torch.from_numpy(value)
-        elif isinstance(value, list) and value:
-            ref_code = torch.as_tensor(value, dtype=torch.long)
-        else:
-            return None
-        if ref_code.ndim == 3:
-            ref_code = ref_code[0]
-        if ref_code.ndim != 2 or ref_code.numel() == 0:
-            return None
-        return ref_code.to(device=device, dtype=torch.long).contiguous()
-
-    @staticmethod
-    def _coerce_token_ids(value: object, *, device: torch.device) -> torch.Tensor | None:
-        value = Qwen3TTSTalkerForConditionalGeneration._first_value(value)
-        if value is None:
-            return None
-        if isinstance(value, torch.Tensor):
-            ids = value
-        elif isinstance(value, np.ndarray):
-            ids = torch.from_numpy(value)
-        elif isinstance(value, list) and value and all(isinstance(v, (int, np.integer)) for v in value):
-            ids = torch.tensor(value, dtype=torch.long)
-        else:
-            return None
-        if ids.ndim == 1:
-            ids = ids.unsqueeze(0)
-        if ids.ndim != 2 or ids.numel() == 0:
-            return None
-        return ids.to(device=device, dtype=torch.long).contiguous()
-
-    @staticmethod
-    def _voice_clone_prompt_dict(raw: object) -> dict[str, object] | None:
-        raw = Qwen3TTSTalkerForConditionalGeneration._first_value(raw)
-        if isinstance(raw, dict):
-            return raw
-        if isinstance(raw, list) and raw and isinstance(raw[0], dict):
-            return raw[0]
-        return None
-
-    @staticmethod
-    def _has_ref_code_like(value: object) -> bool:
-        value = Qwen3TTSTalkerForConditionalGeneration._first_value(value)
-        if isinstance(value, torch.Tensor):
-            return value.numel() > 0 and value.ndim in (2, 3)
-        if isinstance(value, np.ndarray):
-            return value.size > 0 and value.ndim in (2, 3)
-        if isinstance(value, list):
-            return bool(value)
-        return False
-
-    @staticmethod
-    def _needs_initial_prompt_preprocess(info_dict: dict[str, Any]) -> bool:
-        embed = info_dict.get("embed")
-        return not (isinstance(embed, dict) and "prefill" in embed)
-
-    def _needs_batched_ref_code(self, info_dict: dict[str, Any]) -> bool:
-        if not self._needs_initial_prompt_preprocess(info_dict):
-            return False
-        if self._first_value(info_dict.get("task_type"), "CustomVoice") != "Base":
-            return False
-        codes = info_dict.get("codes")
-        if isinstance(codes, dict) and self._has_ref_code_like(codes.get(_PRECOMPUTED_REF_CODE_KEY)):
-            return False
-
-        voice_clone_prompt = self._voice_clone_prompt_dict(info_dict.get("voice_clone_prompt"))
-        if voice_clone_prompt is not None:
-            if self._has_ref_code_like(voice_clone_prompt.get("ref_code")):
-                return False
-            icl_flag = self._first_value(voice_clone_prompt.get("icl_mode"))
-            if isinstance(icl_flag, bool) and not icl_flag:
-                return False
-
-        xvec_only = self._first_value(info_dict.get("x_vector_only_mode"), False)
-        if isinstance(xvec_only, bool) and xvec_only:
-            return False
-
-        ref_audio_list = info_dict.get("ref_audio")
-        return isinstance(ref_audio_list, list) and bool(ref_audio_list)
-
-    @torch.inference_mode()
-    def preprocess_batch(
-        self,
-        *,
-        req_ids: list[str],
-        model_intermediate_buffer: dict[str, dict[str, Any]],
-        device: torch.device,
-    ) -> None:
-        """Batch Base voice-clone ref-audio codec extraction for current prefill requests."""
-        pending_text: list[tuple[dict[str, Any], str]] = []
-        pending_ref_text: list[tuple[dict[str, Any], str]] = []
-        groups: dict[int, list[tuple[dict[str, Any], np.ndarray, int]]] = {}
-        for req_id in req_ids:
-            info_dict = model_intermediate_buffer.get(req_id)
-            if not isinstance(info_dict, dict):
-                continue
-
-            if (
-                self._needs_initial_prompt_preprocess(info_dict)
-                and self._first_value(info_dict.get("task_type"), "CustomVoice") == "Base"
-            ):
-                if _PRECOMPUTED_TEXT_IDS_KEY not in info_dict:
-                    text = self._first_value(info_dict.get("text"), "")
-                    if isinstance(text, str):
-                        pending_text.append((info_dict, self._build_assistant_text(text)))
-
-                if _PRECOMPUTED_REF_IDS_KEY not in info_dict:
-                    ref_text = self._first_value(info_dict.get("ref_text"))
-                    if isinstance(ref_text, str) and ref_text.strip():
-                        pending_ref_text.append((info_dict, self._build_ref_text(ref_text)))
-
-            if not self._needs_batched_ref_code(info_dict):
-                continue
-            ref_audio_list = info_dict.get("ref_audio")
-            if not isinstance(ref_audio_list, list) or not ref_audio_list:
-                continue
-            try:
-                wav, sr = self._normalize_ref_audio(ref_audio_list[0])
-            except Exception:
-                # Keep the original per-request path responsible for surfacing
-                # invalid ref_audio errors with its existing messages.
-                continue
-            info_dict[_NORMALIZED_REF_AUDIO_KEY] = (wav, sr)
-            cache_key = self._make_ref_audio_cache_key(wav, sr)
-            info_dict[_REF_AUDIO_CACHE_KEY] = cache_key
-            cached = self._get_ref_audio_artifacts(cache_key)
-            if cached is not None:
-                cached_ref_code = self._coerce_ref_code_tensor(cached.get("ref_code"), device=device)
-                if isinstance(cached_ref_code, torch.Tensor):
-                    info_dict.setdefault("codes", {})[_PRECOMPUTED_REF_CODE_KEY] = cached_ref_code
-                    continue
-            groups.setdefault(int(sr), []).append((info_dict, wav, int(sr)))
-
-        if pending_text or pending_ref_text:
-            tok_text = self._get_tokenizer()
-            for items, key in (
-                (pending_text, _PRECOMPUTED_TEXT_IDS_KEY),
-                (pending_ref_text, _PRECOMPUTED_REF_IDS_KEY),
-            ):
-                if not items:
-                    continue
-                try:
-                    tokenized = tok_text([text for _, text in items], padding=False)
-                except Exception as exc:
-                    logger.debug("Qwen3-TTS batched text tokenization failed; falling back to serial path: %s", exc)
-                    continue
-                input_ids = tokenized.get("input_ids") if isinstance(tokenized, dict) else None
-                if not isinstance(input_ids, list) or len(input_ids) != len(items):
-                    continue
-                for (info_dict, _), ids in zip(items, input_ids, strict=True):
-                    if isinstance(ids, list) and ids:
-                        info_dict[key] = torch.tensor(ids, dtype=torch.long)
-
-        groups = {sr: items for sr, items in groups.items() if len(items) >= 2}
-        if not groups:
-            return
-
-        for sr, items in groups.items():
-            wavs = [wav for _, wav, _ in items]
-            try:
-                ref_codes = self._encode_ref_audio_batch(wavs, int(sr), device=device)
-            except Exception as exc:
-                logger.debug("Qwen3-TTS batched ref_code encode failed; falling back to serial path: %s", exc)
-                continue
-            for (info_dict, wav, item_sr), ref_code in zip(items, ref_codes, strict=True):
-                ref_code_t = self._coerce_ref_code_tensor(ref_code, device=device)
-                if ref_code_t is None:
-                    continue
-                info_dict.setdefault("codes", {})[_PRECOMPUTED_REF_CODE_KEY] = ref_code_t
-                info_dict[_NORMALIZED_REF_AUDIO_KEY] = (wav, item_sr)
-                cache_key = info_dict.get(_REF_AUDIO_CACHE_KEY)
-                if isinstance(cache_key, str) and cache_key:
-                    self._put_ref_audio_artifacts(cache_key, ref_code=ref_code_t)
-
-    def _generate_icl_prompt(
-        self,
-        *,
-        text_id: torch.Tensor,
-        ref_id: torch.Tensor,
-        ref_code: torch.Tensor,
-        tts_pad_embed: torch.Tensor,
-        tts_eos_embed: torch.Tensor,
-        non_streaming_mode: bool,
-    ) -> tuple[torch.Tensor, torch.Tensor]:
-        # Ported from official Qwen3TTSForConditionalGeneration.generate_icl_prompt
-        text_embed = self.text_projection(self.text_embedding(torch.cat([ref_id, text_id], dim=-1)))
-        text_embed = torch.cat([text_embed, tts_eos_embed], dim=1)
-
-        # codec embed (codec bos + codec) 1 T2 D
-        codec_embed: list[torch.Tensor] = []
-        for i in range(int(self.talker_config.num_code_groups)):
-            if i == 0:
-                codec_embed.append(self.embed_input_ids(ref_code[:, :1]))
-            else:
-                codec_embed.append(self.code_predictor.get_input_embeddings()[i - 1](ref_code[:, i : i + 1]))
-        codec_embed_sum = torch.cat(codec_embed, dim=1).sum(1).unsqueeze(0)  # [1,T,H]
-        codec_embed_sum = torch.cat(
-            [
-                self.embed_input_ids(
-                    torch.tensor([[self.talker_config.codec_bos_id]], device=codec_embed_sum.device, dtype=torch.long)
-                ),
-                codec_embed_sum,
-            ],
-            dim=1,
-        )
-
-        text_lens = int(text_embed.shape[1])
-        codec_lens = int(codec_embed_sum.shape[1])
-        if non_streaming_mode:
-            # Official non-streaming mode: append the full text conditioning in
-            # prefill, and use PAD in decode steps.
-            icl_input_embed = text_embed + self.embed_input_ids(
-                torch.tensor(
-                    [[self.talker_config.codec_pad_id] * text_lens],
-                    device=codec_embed_sum.device,
-                    dtype=torch.long,
-                )
-            )
-            icl_input_embed = torch.cat([icl_input_embed, codec_embed_sum + tts_pad_embed], dim=1)
-            return icl_input_embed, tts_pad_embed
-        if text_lens > codec_lens:
-            return text_embed[:, :codec_lens] + codec_embed_sum, text_embed[:, codec_lens:]
-        text_embed = torch.cat([text_embed] + [tts_pad_embed] * (codec_lens - text_lens), dim=1)
-        return text_embed + codec_embed_sum, tts_pad_embed
-
-    def _build_prompt_embeds(
-        self,
-        *,
-        task_type: str,
-        info_dict: dict[str, Any],
-    ) -> tuple[torch.Tensor, torch.Tensor, int | None, torch.Tensor | None]:
-        text = (info_dict.get("text") or [""])[0]
-        language = (info_dict.get("language") or ["Auto"])[0]
-        non_streaming_mode_val = info_dict.get("non_streaming_mode")
-        if isinstance(non_streaming_mode_val, list):
-            non_streaming_mode_raw = non_streaming_mode_val[0] if non_streaming_mode_val else None
-        else:
-            non_streaming_mode_raw = non_streaming_mode_val
-        if isinstance(non_streaming_mode_raw, bool):
-            non_streaming_mode = non_streaming_mode_raw
-        else:
-            # Match official inference defaults:
-            # - CustomVoice/VoiceDesign: non_streaming_mode=True
-            # - Base: non_streaming_mode=False
-            non_streaming_mode = task_type in ("CustomVoice", "VoiceDesign")
-
-        # Text ids for assistant template (always).
-        tok = self._get_tokenizer()
-        dev = next(self.parameters()).device
-        input_ids = self._coerce_token_ids(info_dict.pop(_PRECOMPUTED_TEXT_IDS_KEY, None), device=dev)
-        if input_ids is None:
-            input_ids = tok(self._build_assistant_text(text), return_tensors="pt", padding=False)["input_ids"].to(
-                device=dev
-            )
-
-        # Optional instruct prefix.
-        instruct = (info_dict.get("instruct") or [""])[0]
-        instruct_embed = None
-        if isinstance(instruct, str) and instruct.strip():
-            instruct_ids = tok(self._build_instruct_text(instruct), return_tensors="pt", padding=False)["input_ids"].to(
-                device=input_ids.device
-            )
-            instruct_embed = self.text_projection(self.text_embedding(instruct_ids))
-
-        # tts special token embeds (projected into talker hidden).
-        # ``tts_pad_embed`` is precomputed in :meth:`_init_runtime_buffers`
-        # (request-independent), so we only need bos/eos here.
-        tts_tokens = torch.tensor(
-            [[self.config.tts_bos_token_id, self.config.tts_eos_token_id]],
-            device=input_ids.device,
-            dtype=input_ids.dtype,
-        )
-        tts_bos_embed, tts_eos_embed = self.text_projection(self.text_embedding(tts_tokens)).chunk(2, dim=1)
-        tts_pad_embed = self._tts_pad_embed.to(device=input_ids.device, dtype=tts_bos_embed.dtype).reshape(1, 1, -1)
-
-        # Codec prefill tags.
-        language_id = None
-        if isinstance(language, str) and language.lower() != "auto":
-            language_id = self.talker_config.codec_language_id.get(language.lower())
-        # Match official dialect override:
-        # If language is Chinese/Auto and the selected speaker is a dialect voice,
-        # set language_id to that dialect to improve code generation stability.
-        if language_id is None and isinstance(language, str) and language.lower() in ("chinese", "auto"):
-            speaker_for_dialect = None
-            if task_type == "CustomVoice":
-                speaker_for_dialect = (info_dict.get("speaker") or [""])[0]
-            if isinstance(speaker_for_dialect, str) and speaker_for_dialect.strip():
-                spk_is_dialect = getattr(self.talker_config, "spk_is_dialect", None) or {}
-                dialect = spk_is_dialect.get(speaker_for_dialect.lower())
-                if isinstance(dialect, str) and dialect:
-                    language_id = self.talker_config.codec_language_id.get(dialect)
-        if language_id is None:
-            codec_prefill_list = [
-                [
-                    self.talker_config.codec_nothink_id,
-                    self.talker_config.codec_think_bos_id,
-                    self.talker_config.codec_think_eos_id,
-                ]
-            ]
-        else:
-            codec_prefill_list = [
-                [
-                    self.talker_config.codec_think_id,
-                    self.talker_config.codec_think_bos_id,
-                    int(language_id),
-                    self.talker_config.codec_think_eos_id,
-                ]
-            ]
-
-        codec_input_0 = self.embed_input_ids(
-            torch.tensor(codec_prefill_list, device=input_ids.device, dtype=torch.long)
-        )
-        codec_input_1 = self.embed_input_ids(
-            torch.tensor([[self.talker_config.codec_pad_id, self.talker_config.codec_bos_id]], device=input_ids.device)
-        )
-
-        # Speaker embedding/token (task-dependent)
-        speaker_embed = None
-        ref_code_len: int | None = None
-        ref_code_prompt: torch.Tensor | None = None
-
-        def _as_singleton(x: object) -> object:
-            if isinstance(x, list):
-                return x[0] if x else None
-            return x
-
-        def _to_long_tensor(x: object, *, device: torch.device) -> torch.Tensor | None:
-            x = _as_singleton(x)
-            if x is None:
-                return None
-            if isinstance(x, torch.Tensor):
-                t = x
-            elif isinstance(x, np.ndarray):
-                t = torch.from_numpy(x)
-            elif isinstance(x, list) and x and all(isinstance(v, (int, np.integer)) for v in x):
-                t = torch.tensor(x, dtype=torch.long)
-            else:
-                return None
-            if t.ndim == 1:
-                t = t.unsqueeze(0)
-            return t.to(device=device, dtype=torch.long)
-
-        def _normalize_voice_clone_prompt(raw: object) -> dict[str, object] | None:
-            raw = _as_singleton(raw)
-            if raw is None:
-                return None
-            if isinstance(raw, dict):
-                return raw
-            # Some callers may pass list[dict] directly.
-            if isinstance(raw, list) and raw and isinstance(raw[0], dict):
-                return raw[0]
-            return None
-
-        if task_type == "Base":
-            # Base supports voice clone prompt with in-context mode.
-            xvec_only = bool((info_dict.get("x_vector_only_mode") or [False])[0])
-            in_context_mode = not xvec_only
-            voice_clone_prompt = _normalize_voice_clone_prompt(info_dict.get("voice_clone_prompt"))
-            ref_audio_wav: np.ndarray | None = None
-            ref_audio_sr: int | None = None
-            ref_audio_cache_key: str | None = None
-
-            def _get_ref_audio() -> tuple[np.ndarray, int]:
-                nonlocal ref_audio_wav, ref_audio_sr
-                if ref_audio_wav is None or ref_audio_sr is None:
-                    normalized_ref_audio = info_dict.pop(_NORMALIZED_REF_AUDIO_KEY, None)
-                    if (
-                        isinstance(normalized_ref_audio, tuple)
-                        and len(normalized_ref_audio) == 2
-                        and isinstance(normalized_ref_audio[0], np.ndarray)
-                    ):
-                        ref_audio_wav = normalized_ref_audio[0]
-                        ref_audio_sr = int(normalized_ref_audio[1])
-                        return ref_audio_wav, ref_audio_sr
-                    ref_audio_list = info_dict.get("ref_audio")
-                    if not isinstance(ref_audio_list, list) or not ref_audio_list:
-                        raise ValueError("Base requires `ref_audio`.")
-                    ref_audio_wav, ref_audio_sr = self._normalize_ref_audio(ref_audio_list[0])
-                return ref_audio_wav, ref_audio_sr
-
-            def _get_ref_audio_cache_key() -> str:
-                nonlocal ref_audio_cache_key
-                if isinstance(ref_audio_cache_key, str) and ref_audio_cache_key:
-                    return ref_audio_cache_key
-                cached_key = info_dict.get(_REF_AUDIO_CACHE_KEY)
-                if isinstance(cached_key, str) and cached_key:
-                    ref_audio_cache_key = cached_key
-                    return ref_audio_cache_key
-                wav_np, sr = _get_ref_audio()
-                ref_audio_cache_key = self._make_ref_audio_cache_key(wav_np, sr)
-                info_dict[_REF_AUDIO_CACHE_KEY] = ref_audio_cache_key
-                return ref_audio_cache_key
-
-            def _try_get_ref_audio_cache_key() -> str | None:
-                nonlocal ref_audio_cache_key
-                if isinstance(ref_audio_cache_key, str) and ref_audio_cache_key:
-                    return ref_audio_cache_key
-                cached_key = info_dict.get(_REF_AUDIO_CACHE_KEY)
-                if isinstance(cached_key, str) and cached_key:
-                    ref_audio_cache_key = cached_key
-                    return ref_audio_cache_key
-                ref_audio_list = info_dict.get("ref_audio")
-                if isinstance(ref_audio_list, list) and ref_audio_list:
-                    return _get_ref_audio_cache_key()
-                return None
-
-            # Speaker cache: only for uploaded (named) speakers
-            _speaker_cache_key = None
-            if voice_clone_prompt is None:
-                _speaker_list = info_dict.get("speaker")
-                if isinstance(_speaker_list, list) and _speaker_list:
-                    _voice_name = str(_speaker_list[0]).lower()
-                    # Per-mode namespace — xvec and icl produce different artifacts
-                    # for the same voice, so they must not share a cache slot.
-                    _mode = "xvec" if xvec_only else "icl"
-                    _voice_created_at = int((info_dict.get("voice_created_at") or [0])[0])
-                    _speaker_cache_key = self._speaker_cache.make_cache_key(
-                        _voice_name,
-                        model_type=f"qwen3_tts_{_mode}",
-                        created_at=_voice_created_at,
-                    )
-                    _cached = self._speaker_cache.get(_speaker_cache_key)
-                    if _cached is not None:
-                        # Transfer cached tensors to current device
-                        ref_code_cached = _cached.get("ref_code")
-                        ref_spk_embed_cached = _cached.get("ref_spk_embedding")
-                        if isinstance(ref_code_cached, torch.Tensor):
-                            ref_code_cached = ref_code_cached.to(device=input_ids.device)
-                        if isinstance(ref_spk_embed_cached, torch.Tensor):
-                            ref_spk_embed_cached = ref_spk_embed_cached.to(device=input_ids.device)
-                        voice_clone_prompt = {
-                            "ref_code": ref_code_cached,
-                            "ref_spk_embedding": ref_spk_embed_cached,
-                            "icl_mode": _cached.get("icl_mode"),
-                        }
-                        _speaker_cache_key = None  # hit → don't store again
-
-            # Official implementation may pass `voice_clone_prompt.icl_mode`.
-            if voice_clone_prompt is not None and "icl_mode" in voice_clone_prompt:
-                icl_flag = _as_singleton(voice_clone_prompt.get("icl_mode"))
-                if isinstance(icl_flag, bool):
-                    in_context_mode = icl_flag
-                    xvec_only = not in_context_mode
-            ref_code = None
-            if voice_clone_prompt is not None:
-                ref_code = _as_singleton(voice_clone_prompt.get("ref_code"))
-            ref_code_t = None
-            if isinstance(ref_code, torch.Tensor):
-                ref_code_t = ref_code
-            elif isinstance(ref_code, np.ndarray):
-                ref_code_t = torch.from_numpy(ref_code)
-            if isinstance(ref_code_t, torch.Tensor):
-                if ref_code_t.ndim == 3:
-                    ref_code_t = ref_code_t[0]
-                ref_code_t = ref_code_t.to(device=input_ids.device, dtype=torch.long)
-                ref_code_len = int(ref_code_t.shape[0])
-            else:
-                codes = info_dict.get("codes")
-                precomputed_ref_code = codes.get(_PRECOMPUTED_REF_CODE_KEY) if isinstance(codes, dict) else None
-                ref_code_t = self._coerce_ref_code_tensor(precomputed_ref_code, device=input_ids.device)
-                if isinstance(ref_code_t, torch.Tensor):
-                    ref_code_len = int(ref_code_t.shape[0])
-            if ref_code_t is None and in_context_mode:
-                cache_key = _get_ref_audio_cache_key()
-                cached = self._get_ref_audio_artifacts(cache_key)
-                cached_ref_code = None
-                if isinstance(cached, dict):
-                    cached_ref_code = cached.get("ref_code")
-                ref_code_t = self._coerce_ref_code_tensor(cached_ref_code, device=input_ids.device)
-                if isinstance(ref_code_t, torch.Tensor):
-                    ref_code_len = int(ref_code_t.shape[0])
-                else:
-                    wav_np, sr = _get_ref_audio()
-                    ref_code_t = self._encode_ref_audio_batch([wav_np], int(sr), device=input_ids.device)[0]
-                    ref_code_len = int(ref_code_t.shape[0])
-                    self._put_ref_audio_artifacts(cache_key, ref_code=ref_code_t)
-            if isinstance(ref_code_t, torch.Tensor):
-                ref_code_prompt = ref_code_t
-                cache_key = _try_get_ref_audio_cache_key()
-                if isinstance(cache_key, str) and cache_key:
-                    self._put_ref_audio_artifacts(cache_key, ref_code=ref_code_prompt)
-
-            # Speaker embedding: use prompt embed if provided; otherwise extract from audio.
-            # NOTE: Do NOT use _as_singleton here — the embedding may be a plain
-            # float list (from API via msgspec IPC) that _as_singleton would
-            # destructively unwrap to a single scalar.
-            spk = None
-            if voice_clone_prompt is not None:
-                spk = voice_clone_prompt.get("ref_spk_embedding")
-            if isinstance(spk, torch.Tensor):
-                speaker_embed = spk.to(device=input_ids.device, dtype=torch.bfloat16).view(1, 1, -1)
-            elif isinstance(spk, (list, np.ndarray)):
-                # Plain list/array from API (survived msgspec IPC serialization).
-                speaker_embed = torch.tensor(spk, dtype=torch.bfloat16, device=input_ids.device).view(1, 1, -1)
-            else:
-                cache_key = _get_ref_audio_cache_key()
-                cached = self._get_ref_audio_artifacts(cache_key)
-                cached_spk = None
-                if isinstance(cached, dict):
-                    cached_spk = cached.get("ref_spk_embedding")
-                if isinstance(cached_spk, torch.Tensor):
-                    speaker_embed = cached_spk.to(device=input_ids.device, dtype=torch.bfloat16).view(1, 1, -1)
-                else:
-                    wav_np, sr = _get_ref_audio()
-                    speaker_embed = self._extract_speaker_embedding(wav_np, sr).view(1, 1, -1)
-                    self._put_ref_audio_artifacts(cache_key, ref_spk_embedding=speaker_embed)
-
-            # Cache miss: store extraction result
-            if _speaker_cache_key is not None and speaker_embed is not None:
-                self._speaker_cache.put(
-                    _speaker_cache_key,
-                    {
-                        "ref_code": ref_code_prompt.detach().cpu()
-                        if isinstance(ref_code_prompt, torch.Tensor)
-                        else None,
-                        "ref_spk_embedding": speaker_embed.detach().cpu().reshape(-1),
-                        "icl_mode": in_context_mode,
-                    },
-                )
-
-            codec_input = torch.cat([codec_input_0, speaker_embed, codec_input_1], dim=1)
-
-            # Role header (<|im_start|>assistant\n) -> projected text embeds.
-            role_embed = self.text_projection(self.text_embedding(input_ids[:, :3]))
-
-            codec_prefix = torch.cat((tts_pad_embed.expand(-1, codec_input.shape[1] - 2, -1), tts_bos_embed), dim=1)
-            codec_prefix = codec_prefix + codec_input[:, :-1]
-            talker_prompt = torch.cat((role_embed, codec_prefix), dim=1)
-
-            if in_context_mode:
-                # Prefer explicit tokenized `ref_ids` if provided (matches official signature).
-                ref_ids = self._coerce_token_ids(info_dict.pop(_PRECOMPUTED_REF_IDS_KEY, None), device=input_ids.device)
-                if ref_ids is None:
-                    ref_ids = _to_long_tensor(info_dict.get("ref_ids"), device=input_ids.device)
-                if ref_ids is None and voice_clone_prompt is not None:
-                    ref_ids = _to_long_tensor(
-                        voice_clone_prompt.get("ref_ids") or voice_clone_prompt.get("ref_id"), device=input_ids.device
-                    )
-                if ref_ids is None:
-                    ref_text = _as_singleton(info_dict.get("ref_text"))
-                    if isinstance(ref_text, str) and ref_text.strip():
-                        ref_ids = tok(
-                            self._build_ref_text(ref_text),
-                            return_tensors="pt",
-                            padding=False,
-                        )["input_ids"].to(device=input_ids.device)
-                    else:
-                        logger.warning("Base ICL: ref_text/ref_ids missing, falling back to x-vector-only mode.")
-                        in_context_mode = False
-            if in_context_mode:
-                icl_input_embed, trailing_text_hidden = self._generate_icl_prompt(
-                    text_id=input_ids[:, 3:-5],
-                    ref_id=ref_ids[:, 3:-2],
-                    ref_code=ref_code_t,  # type: ignore[arg-type]
-                    tts_pad_embed=tts_pad_embed,
-                    tts_eos_embed=tts_eos_embed,
-                    non_streaming_mode=non_streaming_mode,
-                )
-                talker_prompt = torch.cat([talker_prompt, icl_input_embed], dim=1)
-            else:
-                # First text token (+ codec_bos).
-                if non_streaming_mode:
-                    # Official non-streaming mode: put the full text into the
-                    # prefill prompt and use PAD for decode steps.
-                    text_all = self.text_projection(self.text_embedding(input_ids[:, 3:-5]))
-                    text_all = torch.cat([text_all, tts_eos_embed], dim=1)
-                    pad_ids = torch.full(
-                        (1, int(text_all.shape[1])),
-                        int(self.talker_config.codec_pad_id),
-                        device=input_ids.device,
-                        dtype=torch.long,
-                    )
-                    talker_prompt = torch.cat(
-                        [
-                            talker_prompt,
-                            text_all + self.embed_input_ids(pad_ids),
-                            tts_pad_embed
-                            + self.embed_input_ids(
-                                torch.tensor([[self.talker_config.codec_bos_id]], device=input_ids.device)
-                            ),
-                        ],
-                        dim=1,
-                    )
-                    trailing_text_hidden = tts_pad_embed
-                else:
-                    first_text = self.text_projection(self.text_embedding(input_ids[:, 3:4])) + codec_input[:, -1:]
-                    talker_prompt = torch.cat([talker_prompt, first_text], dim=1)
-                    trailing_text_hidden = torch.cat(
-                        (
-                            self.text_projection(self.text_embedding(input_ids[:, 4:-5])),
-                            tts_eos_embed,
-                        ),
-                        dim=1,
-                    )
-
-        elif task_type == "CustomVoice":
-            _speaker_raw = info_dict.get("speaker") or [""]
-            speaker = (
-                ((_speaker_raw[0] if isinstance(_speaker_raw, (list, tuple)) else _speaker_raw) or "").lower().strip()
-            )
-            if not speaker:
-                raise ValueError("CustomVoice requires additional_information.speaker.")
-            spk_id_map = {k.lower(): v for k, v in (getattr(self.talker_config, "spk_id", None) or {}).items()}
-            if speaker not in spk_id_map:
-                raise ValueError(f"Unsupported speaker: {speaker}")
-            spk_id = spk_id_map[speaker]
-            # Keep it at least 1D; embedding on a 0-d tensor can return 1D.
-            spk_tensor = torch.tensor([spk_id], device=input_ids.device, dtype=torch.long)
-            spk_embed = self.embed_input_ids(spk_tensor)
-            if spk_embed.ndim == 1:
-                spk_embed = spk_embed.view(1, 1, -1)
-            elif spk_embed.ndim == 2:
-                spk_embed = spk_embed.view(1, 1, -1)
-            speaker_embed = spk_embed
-            codec_input = torch.cat([codec_input_0, speaker_embed, codec_input_1], dim=1)
-
-            role_embed = self.text_projection(self.text_embedding(input_ids[:, :3]))
-            codec_prefix = torch.cat((tts_pad_embed.expand(-1, codec_input.shape[1] - 2, -1), tts_bos_embed), dim=1)
-            codec_prefix = codec_prefix + codec_input[:, :-1]
-            talker_prompt = torch.cat((role_embed, codec_prefix), dim=1)
-
-            if non_streaming_mode:
-                text_all = self.text_projection(self.text_embedding(input_ids[:, 3:-5]))
-                text_all = torch.cat([text_all, tts_eos_embed], dim=1)
-                pad_ids = torch.full(
-                    (1, int(text_all.shape[1])),
-                    int(self.talker_config.codec_pad_id),
-                    device=input_ids.device,
-                    dtype=torch.long,
-                )
-                talker_prompt = torch.cat(
-                    [
-                        talker_prompt,
-                        text_all + self.embed_input_ids(pad_ids),
-                        tts_pad_embed
-                        + self.embed_input_ids(
-                            torch.tensor([[self.talker_config.codec_bos_id]], device=input_ids.device)
-                        ),
-                    ],
-                    dim=1,
-                )
-                trailing_text_hidden = tts_pad_embed
-            else:
-                first_text = self.text_projection(self.text_embedding(input_ids[:, 3:4])) + codec_input[:, -1:]
-                talker_prompt = torch.cat([talker_prompt, first_text], dim=1)
-                trailing_text_hidden = torch.cat(
-                    (
-                        self.text_projection(self.text_embedding(input_ids[:, 4:-5])),
-                        tts_eos_embed,
-                    ),
-                    dim=1,
-                )
-
-        elif task_type == "VoiceDesign":
-            # No known speaker identity; only codec tags + text.
-            codec_input = torch.cat([codec_input_0, codec_input_1], dim=1)
-
-            role_embed = self.text_projection(self.text_embedding(input_ids[:, :3]))
-            codec_prefix = torch.cat((tts_pad_embed.expand(-1, codec_input.shape[1] - 2, -1), tts_bos_embed), dim=1)
-            codec_prefix = codec_prefix + codec_input[:, :-1]
-            talker_prompt = torch.cat((role_embed, codec_prefix), dim=1)
-
-            if non_streaming_mode:
-                text_all = self.text_projection(self.text_embedding(input_ids[:, 3:-5]))
-                text_all = torch.cat([text_all, tts_eos_embed], dim=1)
-                pad_ids = torch.full(
-                    (1, int(text_all.shape[1])),
-                    int(self.talker_config.codec_pad_id),
-                    device=input_ids.device,
-                    dtype=torch.long,
-                )
-                talker_prompt = torch.cat(
-                    [
-                        talker_prompt,
-                        text_all + self.embed_input_ids(pad_ids),
-                        tts_pad_embed
-                        + self.embed_input_ids(
-                            torch.tensor([[self.talker_config.codec_bos_id]], device=input_ids.device)
-                        ),
-                    ],
-                    dim=1,
-                )
-                trailing_text_hidden = tts_pad_embed
-            else:
-                first_text = self.text_projection(self.text_embedding(input_ids[:, 3:4])) + codec_input[:, -1:]
-                talker_prompt = torch.cat([talker_prompt, first_text], dim=1)
-                trailing_text_hidden = torch.cat(
-                    (
-                        self.text_projection(self.text_embedding(input_ids[:, 4:-5])),
-                        tts_eos_embed,
-                    ),
-                    dim=1,
-                )
-        else:
-            raise ValueError(f"Unsupported task_type={task_type}")
-
-        if instruct_embed is not None:
-            talker_prompt = torch.cat([instruct_embed, talker_prompt], dim=1)
-
-        return (
-            talker_prompt.squeeze(0),  # [prompt_len, H]
-            trailing_text_hidden.squeeze(0),  # [T, H]
-            ref_code_len,
-            ref_code_prompt.contiguous() if isinstance(ref_code_prompt, torch.Tensor) else None,
-        )
 
     def load_weights(self, weights: Iterable[tuple[str, torch.Tensor]]) -> set[str]:
         # Consume talker weights, and conditionally consume speaker encoder
@@ -2253,17 +958,6 @@ class Qwen3TTSTalkerForConditionalGeneration(nn.Module):
         self._stacked_codec_embed = torch.stack([e.weight.detach() for e in embeds], dim=0).to(
             device=w.device, dtype=w.dtype
         )
-
-    def _get_resampler(self, orig_sr: int, target_sr: int) -> AudioResampler:
-        key = (orig_sr, target_sr)
-        cache = self._resampler_cache
-        if key in cache:
-            cache.move_to_end(key)
-            return cache[key]
-        cache[key] = AudioResampler(target_sr=target_sr)
-        if len(cache) > self._resampler_cache_max:
-            cache.popitem(last=False)
-        return cache[key]
 
     @torch.no_grad()
     def _init_runtime_buffers(self) -> None:
